Session 1 Steps 5-7: sanity experiment + figures + README update
4-variant experiment on 1000 Materials Project oxide structures:
- Supervised (100%): MAE = 3.78 deg (ceiling)
- Sort-match SSL (10%): MAE = 3.87 deg (98% of supervised!)
- Supervised (10%): MAE = 7.11 deg (limited labels only)
- Random match: MAE = 11.97 deg (broken baseline)

Key findings:
- SSL with 10% labels achieves 98% of fully supervised performance
- SSL cuts error by 46% vs supervised-10% (7.11 -> 3.87)
- Random matching is 3x worse, validating the theorem
- 20 epochs, 257k params, ~11s per variant on CPU

Figures: training curves (4 variants) + predicted vs true scatter

+++ src/model.py
@@ -136,8 +136,8 @@ class XRDPredictor(nn.Module):
         Returns:
             pred_2theta: (B, n_peaks) predicted peak positions in degrees.
         """
-        x = self.node_embed(graph.node_feats)        # (N, hidden)
-        edge_feats = self.edge_embed(graph.edge_feats)  # (E, hidden)
+        x = self.node_embed(graph.node_feats.float())        # (N, hidden)
+        edge_feats = self.edge_embed(graph.edge_feats.float())  # (E, hidden)
 
         # Message passing
         for gin in self.gin_layers:
